Added a state machine
Use xstate to provide a very simple state machine. Chosen state to manage is online/offline status, which needs a subscription object to be included in order to organise the event handlers. In principle you don't need a state machine for this, but you can see the illustration of it's impact really in the 'actions' function updateStatus, where we don't need to know the outcome of changing status - this is handled by the state machine.